requirements/base.txt
Django==1.7
certifi==14.05.14
django-grappelli==2.6.2
djangorestframework==2.4.3
psycopg2==2.5.4
requests==2.4.3
tornado==4.0.2
